requirements.txt
blinker==1.7.0
click==8.1.7
distlib==0.3.7
dominate==2.9.1
filelock==3.13.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
platformdirs==4.0.0
python-bitcoinlib==0.12.2
virtualenv==20.25.0
visitor==0.1.3
Werkzeug==3.0.1
